fix(daemon): replicate bare-verb auto-route to capture
The CLI fast path in core/index.ts auto-routes any unknown verb to
`prjct capture "<verb plus args>"` so dumps without ceremony land in the
GTD inbox. The daemon path was missing that fallback — unknown verbs
fell through to the registry default and returned "Command not found"
instead.

Symptom (pre-fix):
  $ prjct "another inbox item"
  Command not found: another

(Worked only with PRJCT_NO_DAEMON=1, which bypasses the daemon.)

Fix:
- Extract `findClosestCommand` (+ its `editDistance` helper) from
  core/index.ts into a new `core/commands/closest-command.ts` so both
  entry points can share the typo guard without duplicating Levenshtein.
- In `core/daemon/daemon.ts:executeCommand()`, mirror the auto-route
  block from core/index.ts: unknown verb + not a single-word near-typo
  → route to `commands.capture(...)`.

Tests: 866/866 still pass. typecheck + check + build all green.

Detected end-to-end on v2.2.8 install during the post-#256 e2e sweep.
