+++ core/index.ts
@@ -4,6 +4,7 @@
  * This file is required by bin/prjct after setup verification
  */
 
+import { findClosestCommand } from './commands/closest-command'
 import { PrjctCommands } from './commands/commands'
 import { commandRegistry } from './commands/registry'
 import './commands/register' // Ensure commands are registered
@@ -324,44 +325,6 @@ function validateCommandParams(
   }
 
   return null
-}
-
-/**
- * Find closest matching command name for did-you-mean suggestions
- * Uses Levenshtein edit distance — suggests if distance <= 2
- */
-function findClosestCommand(input: string): string | null {
-  const allNames = commandRegistry.getAll().map((c) => c.name)
-  let best: string | null = null
-  let bestDist = Infinity
-
-  for (const name of allNames) {
-    const dist = editDistance(input.toLowerCase(), name.toLowerCase())
-    if (dist < bestDist) {
-      bestDist = dist
-      best = name
-    }
-  }
-
-  // Only suggest if edit distance is at most 2
-  return bestDist <= 2 ? best : null
-}
-
-function editDistance(a: string, b: string): number {
-  const m = a.length
-  const n = b.length
-  const dp: number[][] = Array.from({ length: m + 1 }, () => Array(n + 1).fill(0))
-  for (let i = 0; i <= m; i++) dp[i][0] = i
-  for (let j = 0; j <= n; j++) dp[0][j] = j
-  for (let i = 1; i <= m; i++) {
-    for (let j = 1; j <= n; j++) {
-      dp[i][j] =
-        a[i - 1] === b[j - 1]
-          ? dp[i - 1][j - 1]
-          : 1 + Math.min(dp[i - 1][j], dp[i][j - 1], dp[i - 1][j - 1])
-    }
-  }
-  return dp[m][n]
 }
 
 /**
